Reorganization and improved products from collection component
