+++ core/belami/components/product/index.tsx
@@ -1,0 +1,4 @@
+export { CollectionProducts } from './collection-products';
+//export { CollectionProductsStatic } from './collection-products-static';
+export { RelatedProducts } from './related-products';
+//export { SimilarProducts } from './similar-products';
