+++ core/belami/components/product/related-products.tsx
@@ -2,21 +2,20 @@
 
 import React from 'react';
 import { useState, useEffect } from 'react';
-
 import Link from 'next/link';
 import Image from 'next/image';
-//import { BcImage } from '~/components/bc-image';
-//import { imageManagerImageUrl } from '~/lib/store-assets';
 
 import noImage from '~/public/no-image.svg';
-import blurredRectangle from '~/public/other/blurred-rectangle.jpg';
 
-//import { Carousel } from '~/components/ui/carousel2';
-//import { ProductCard } from '~/components/product-card';
+import {
+  Carousel,
+  CarouselButtons,
+  CarouselContent,
+  CarouselItem,
+  CarouselScrollbar,
+} from '~/belami/components/carousel';
 
 import { useFormatter } from 'next-intl';
-
-import { cn } from '~/lib/utils';
 
 import { ReviewSummary } from '~/belami/components/reviews';
 
@@ -30,14 +29,11 @@ const searchColorsHEX: DynamicObject = searchColors;
 
 //const useDefaultPrices = process.env.NEXT_PUBLIC_USE_DEFAULT_PRICES === 'true';
 const useAsyncMode = process.env.NEXT_PUBLIC_USE_ASYNC_MODE === 'true';
-//const useAsyncMode = false;
 
 interface Props {
-  collection: string,
+  productId: number,
   products: any[],
-  total?: number,
-  useDefaultPrices?: boolean,
-  moreLink?: string | null
+  useDefaultPrices?: boolean
 }
 
 function getDiscount(price: number, sale: number): number | null {
@@ -69,7 +65,7 @@ function ColorSwatches({ variants, onImageClick }: any) {
   )
 }
 
-function CustomItem({ hit, useDefaultPrices = false, price = null, salePrice = null, isLoading = false, isLoaded = false, className = {} }: any) {
+function CustomItem({ hit, useDefaultPrices = false, price = null, salePrice = null, isLoading = false, isLoaded = false }: any) {
 
   const format = useFormatter();
   const currency = 'USD';
@@ -81,8 +77,8 @@ function CustomItem({ hit, useDefaultPrices = false, price = null, salePrice = n
   }
 
   return (
-    <article data-id={hit.objectID} className={cn("w-full product flex flex-col h-full", className)}>
-      <div>
+    <article data-id={hit.objectID} className="w-full product flex flex-col h-full border border-gray-300 rounded-none">
+      <div className="px-4 pt-4">
         <div className="flex my-0 mx-auto pb-full w-full h-auto overflow-hidden pb-[100%] relative">
           <figure className="absolute left-0 top-0 w-full h-full">
             <Link href={hit.url} className="w-full h-full flex justify-center items-center align-middle bg-white">
@@ -139,34 +135,20 @@ function CustomItem({ hit, useDefaultPrices = false, price = null, salePrice = n
             <ReviewSummary numberOfReviews={hit.reviews_count} averageRating={hit.reviews_rating_sum} className="mx-auto mt-2 justify-center" />
           }
         </div>
-      </div>
-    </article>
-  );
-}
-
-
-function MoreItem({ total = 0, moreLink = null, className = {} }: any) {
-  return (total > 0 &&
-    <article className={cn("collection-products-more w-full flex flex-col h-full", className)}>
-      <div>
-        <div className="flex my-0 mx-auto pb-full w-full h-auto overflow-hidden pb-[100%] relative">
-          <figure className="absolute left-0 top-0 w-full h-full">
-            <div className="w-full h-full flex flex-col justify-center items-center align-middle bg-brand-300 bg-no-repeat bg-cover" style={{backgroundImage: `url(${blurredRectangle.src})`}}>
-              <h2 className="flex text-lg font-medium text-white">+ {total} Items</h2>
-              {moreLink && <Link href={moreLink} className="mx-auto mt-1 flex items-center justify-center space-x-2 px-8 h-10 rounded uppercase bg-white text-gray-700">Shop</Link>}
-            </div>
-          </figure>
+        <div className="mt-4 p-4 hidden lg:block">
+          <button type="button" className="flex items-center justify-center space-x-2 px-4 w-full h-10 rounded border border-[#B4DDE9] cursor-pointer uppercase" onClick={(e) => alert('Coming Soon!')}>
+            <svg width="18" height="18" viewBox="0 0 17 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+              <path d="M13 12H12.565L11.9575 11.3925C12.625 10.5 13 9.375 13 8.25C13 5.3475 10.6525 3 7.75 3C6.625 3 5.5 3.375 4.5925 4.05C2.275 5.79 1.8025 9.0825 3.5425 11.4C5.2825 13.7175 8.575 14.19 10.8925 12.45L11.5 13.0575V13.5L15.25 17.25L16.75 15.75L13 12ZM7.75 12C5.68 12 4 10.32 4 8.25C4 6.18 5.68 4.5 7.75 4.5C9.82 4.5 11.5 6.18 11.5 8.25C11.5 10.32 9.82 12 7.75 12ZM1.75 4.5L0.25 6V0.75H5.5L4 2.25H1.75V4.5ZM15.25 0.75V6L13.75 4.5V2.25H11.5L10 0.75H15.25ZM4 14.25L5.5 15.75H0.25V10.5L1.75 12V14.25H4Z" fill="#353535"/>
+            </svg>
+            <span className='font-[500] leading-[32px] text-[14px] tracking-[1.25px] text-[#002A37]'>Quick View</span>
+          </button>
         </div>
       </div>
     </article>
   );
 }
 
-
-
-
-export function CollectionProducts({collection, products, total = 0, useDefaultPrices = false, moreLink = null }: Props) {
-  const breakpoints = ["sm:flex", "md:flex", "lg:flex", "xl:flex", "2xl:flex"];
+export function RelatedProducts({ productId, products, useDefaultPrices = false }: Props) {
 
   const [isLoading, setIsLoading] = useState(false);
   const [isLoaded, setIsLoaded] = useState(false);
@@ -209,12 +191,16 @@ export function CollectionProducts({collection, products, total = 0, useDefaultP
 
   return (
     products && products.length > 0 &&
-    <div className="carousel-container p-4 xl:p-8 bg-gray-50 mt-8 mb-12">
-      <h2 className="text-3xl font-black text-[1.5rem] font-normal leading-[2rem] text-left text-[#353535]">{`More from ${collection} collection`}</h2>
-      <div className="mt-5 grid sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-5 gap-4 xl:gap-8">
-        {products.map((item: any, index: number) => (
+    <div className="mt-4 mb-12">
+      <h2 className="flex space-x-2 items-center text-3xl font-normal leading-[2rem] text-left text-[#353535]">
+        <span className='brightness-0 xl:brightness-[1]'><svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="text-brand-500"><g><path d="M0.974976 7.0001C0.974976 5.33343 1.56248 3.9126 2.73748 2.7376C3.91248 1.5626 5.33331 0.975098 6.99998 0.975098V3.0001C5.89998 3.0001 4.95831 3.39176 4.17498 4.1751C3.39164 4.95843 2.99998 5.9001 2.99998 7.0001H0.974976ZM5.29998 18.7251C3.78331 17.2084 3.02498 15.3834 3.02498 13.2501C3.02498 11.1168 3.78331 9.29176 5.29998 7.7751L7.04998 6.0001L7.34998 6.3001C7.83331 6.78343 8.07498 7.37093 8.07498 8.0626C8.07498 8.75426 7.83331 9.34176 7.34998 9.8251L6.99998 10.1751C6.79998 10.3751 6.69998 10.6126 6.69998 10.8876C6.69998 11.1626 6.79998 11.4001 6.99998 11.6001L7.89998 12.5001C8.33331 12.9334 8.54998 13.4584 8.54998 14.0751C8.54998 14.6918 8.33331 15.2168 7.89998 15.6501L8.97498 16.7251C9.70831 15.9918 10.075 15.1126 10.075 14.0876C10.075 13.0626 9.69998 12.1751 8.94998 11.4251L8.39998 10.8751C8.83331 10.4418 9.14164 9.95426 9.32498 9.4126C9.50831 8.87093 9.58331 8.31676 9.54998 7.7501L14.025 3.2751C14.225 3.0751 14.4625 2.9751 14.7375 2.9751C15.0125 2.9751 15.25 3.0751 15.45 3.2751C15.65 3.4751 15.75 3.7126 15.75 3.9876C15.75 4.2626 15.65 4.5001 15.45 4.7001L10.775 9.3751L11.825 10.4251L17.85 4.4251C18.05 4.2251 18.2833 4.1251 18.55 4.1251C18.8166 4.1251 19.05 4.2251 19.25 4.4251C19.45 4.6251 19.55 4.85843 19.55 5.1251C19.55 5.39176 19.45 5.6251 19.25 5.8251L13.25 11.8501L14.3 12.9001L19.6 7.6001C19.8 7.4001 20.0375 7.3001 20.3125 7.3001C20.5875 7.3001 20.825 7.4001 21.025 7.6001C21.225 7.8001 21.325 8.0376 21.325 8.3126C21.325 8.5876 21.225 8.8251 21.025 9.0251L15.725 14.3251L16.775 15.3751L20.825 11.3251C21.025 11.1251 21.2625 11.0251 21.5375 11.0251C21.8125 11.0251 22.05 11.1251 22.25 11.3251C22.45 11.5251 22.55 11.7626 22.55 12.0376C22.55 12.3126 22.45 12.5501 22.25 12.7501L16.25 18.7251C14.7333 20.2418 12.9083 21.0001 10.775 21.0001C8.64164 21.0001 6.81664 20.2418 5.29998 18.7251ZM17 23.0251V21.0001C18.1 21.0001 19.0416 20.6084 19.825 19.8251C20.6083 19.0418 21 18.1001 21 17.0001H23.025C23.025 18.6668 22.4375 20.0876 21.2625 21.2626C20.0875 22.4376 18.6666 23.0251 17 23.0251Z" fill="currentColor"></path></g></svg></span>
+        <span>You May Also Like ...</span>
+      </h2>
+      <Carousel className="overflow-x-hidden mb-16 mt-5 lg:mt-6">
+        <CarouselContent className="mb-10">
+        {products.map((item: any) => (
+          <CarouselItem className="basis-full sm:basis-1/2 md:basis-1/3 lg:basis-1/3 xl:basis-1/4 2xl:basis-1/5" key={item.objectID}>
           <CustomItem 
-            key={item.objectID} 
             hit={item} 
             useDefaultPrices={useDefaultPrices}
             price={
@@ -229,36 +215,13 @@ export function CollectionProducts({collection, products, total = 0, useDefaultP
             }
             isLoading={isLoading}
             isLoaded={isLoaded}
-            className={cn(
-              index > 0 && `hidden`,
-              index == 1 && total == 0 && `${breakpoints[0]}`,
-              index == 1 && total > 0 && `${breakpoints[1]}`,
-              index == 2 && total == 0 && `${breakpoints[1]}`,
-              index == 2 && total > 0 && `${breakpoints[2]}`,
-              index == 3 && total == 0 && `${breakpoints[2]}`,
-              index == 3 && total > 0 && `${breakpoints[2]}`,
-              index == 4 && total == 0 && `${breakpoints[2]}`,
-              index == 4 && total > 0 && `${breakpoints[2]}`,
-              index == 5 && total == 0 && `${breakpoints[2]}`,
-              index == 5 && total > 0 && `${breakpoints[3]}`,
-              index == 6 && total == 0 && `${breakpoints[3]}`,
-              index == 6 && total > 0 && `${breakpoints[3]}`,
-              index == 7 && total == 0 && `${breakpoints[3]}`,
-              index == 7 && total > 0 && `${breakpoints[4]}`,
-              index == 8 && total == 0 && `${breakpoints[4]}`,
-              index == 8 && total > 0 && `${breakpoints[4]}`,
-              index == 9 && total == 0 && `${breakpoints[4]}`,
-              index == 9 && total > 0 && `${breakpoints[4]}`,
-            )}
-            /*
-            className={cn(
-              index > 0 && index < 6 && `hidden ${breakpoints[index]}` 
-            )}
-            */ 
           />
+          </CarouselItem>
         ))}
-        <MoreItem total={total} moreLink={moreLink} />
-      </div>
+        </CarouselContent>
+        <CarouselScrollbar className="carousel-scrollbar" />
+        <CarouselButtons className="carousel-arrows" />
+      </Carousel>
     </div>
   );
 }
