+++ core/belami/components/product/similar-products.tsx
